requirements.txt
aiogram>=3.13,<4
pandas>=2.2
openpyxl>=3.1
python-dotenv>=1.0
httpx>=0.27,<1
